requirements.txt
streamlit
audio_recorder_streamlit
streamlit_extras
pillow
